extract wizard layout to Wizard component
